Refactor out importers
