+++ AuthenticatorShared/Core/Vault/Services/Importers/Importer.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+// MARK: - Importer
+
+/// A protocol for classes that converts data into items to import.
+protocol Importer {
+    static func importItems(data: Data) throws -> [AuthenticatorItemView]
+}
