+++ AuthenticatorShared/Core/Vault/Services/Importers/RaivoImporter.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+class RaivoImporter: Importer {
+    static func importItems(data: Data) throws -> [AuthenticatorItemView] {
+        return []
+    }
+}
